sample_code/python/vessels_v2_graphql/requirements.txt
gql>=2.0.0
PyYAML>=5.4.1
loguru